add schemas for controllers reporte, email, provedor, pedido, inventario and producto.

+++ backend/src/controllers/pedido.controller.js
@@ -1,8 +1,8 @@
-/* eslint-disable operator-linebreak */
 "use strict";
 import { respondSuccess, respondError } from "../utils/resHandler.js";
 import PedidoService from "../services/pedido.service.js";
 import { handleError } from "../utils/errorHandler.js";
+import { pedidoBodySchema, pedidoIdSchema } from "../schema/pedido.schema.js";
 
 /**
  * Obtiene todos los pedidos
@@ -33,6 +33,9 @@ async function getPedidos(req, res) {
 async function createPedido(req, res) {
     try {
         const { body } = req;
+        const { error: bodyError } = pedidoBodySchema.validate(body);
+        if (bodyError) return respondError(req, res, 400, bodyError.message);
+
         const [newPedido, errorPedido] = await PedidoService.createPedido(body);
         if (errorPedido) return respondError(req, res, 400, errorPedido);
 
@@ -51,6 +54,9 @@ async function createPedido(req, res) {
 async function getPedidoById(req, res) {
     try {
         const { params } = req;
+        const { error: paramsError } = pedidoIdSchema.validate(params);
+        if (paramsError) return respondError(req, res, 400, paramsError.message);
+
         const [pedido, errorPedido] = await PedidoService.getPedidoById(params.id);
         if (errorPedido) return respondError(req, res, 404, errorPedido);
 
@@ -69,6 +75,12 @@ async function getPedidoById(req, res) {
 async function updatePedido(req, res) {
     try {
         const { params, body } = req;
+        const { error: paramsError } = pedidoIdSchema.validate(params);
+        if (paramsError) return respondError(req, res, 400, paramsError.message);
+
+        const { error: bodyError } = pedidoBodySchema.validate(body);
+        if (bodyError) return respondError(req, res, 400, bodyError.message);
+
         const [pedidoUpdated, errorPedido] = await PedidoService.updatePedido(params.id, body);
         if (errorPedido) return respondError(req, res, 400, errorPedido);
 
@@ -87,6 +99,9 @@ async function updatePedido(req, res) {
 async function deletePedido(req, res) {
     try {
         const { params } = req;
+        const { error: paramsError } = pedidoIdSchema.validate(params);
+        if (paramsError) return respondError(req, res, 400, paramsError.message);
+
         const [pedidoDeleted, errorPedido] = await PedidoService.deletePedido(params.id);
         if (errorPedido) return respondError(req, res, 404, errorPedido);
